Query for /aggregateDataset?q=/STAT
